requirements.txt
httpx>=0.27,<1
beautifulsoup4>=4.12,<5
lxml>=5.1,<6
APScheduler>=3.10,<4
python-telegram-bot>=21.0,<22
